requirements.txt
discord.py==2.6.4
python-aternos==3.0.6
flask==3.1.2
